backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
PyJWT==2.9.0
python-multipart==0.0.9
psycopg2-binary==2.9.9

